graph drawing: Initial work on supporting clusters to be defined.
I'm not entirely happy having a /tikz/graphs/cluster key in
tikzlibrarygraphdrawing.code.tex, so maybe we can change that later.

This commit introduces a new macro \pgfgdaddnodetocluster{cluster
name}{node name}. It's name is self-explanatory, I guess. It simply
calls Interface:addNodeToCluster(node_name, cluster_name) which in turn
adds the node to the given cluster of the graph.
The Graph class holds a new clusters array where clusters are stored in
the order they were defined in TikZ. These clusters can be used for
implementing constraints for groups of nodes such as placing all nodes
of a cluster in the same layer of a layered drawing.

This is also what the GansnerKNV1993 node ranking algorithm does now. It
merges all nodes of a cluster into a single node, re-routes edges as
necessary and introduces new edges from cluster N to cluster N+1 in
order to force clusters to be on separate ranks. This is highly
experimental though. The "modular layered" algorithm now has to remove
loops and merge multi-edges two times, once before merging clusters and
once after dissolving clusters again. For this, I added functions in the
manipulation namespace but this is more of a temporary hack.

The Edge class was reworked a bit internally and now has a new
Edge:isLoop() method.